+++ packages/babel-plugin-jest-hoist/src/__tests__/integration-test.js
@@ -17,6 +17,7 @@ import a from '../__test_modules__/a';
 import b from '../__test_modules__/b';
 import c from '../__test_modules__/c';
 import d from '../__test_modules__/d';
+import e from '../__test_modules__/e';
 
 // These will all be hoisted above imports
 jest.unmock('react');
@@ -24,14 +25,30 @@ jest.unmock('../__test_modules__/Unmocked');
 jest
   .unmock('../__test_modules__/c')
   .unmock('../__test_modules__/d');
+jest.mock('../__test_modules__/e', () => {
+  if (!global.CALLS) {
+    global.CALLS = 0;
+  }
+  global.CALLS++;
+
+  return {
+    _isMock: true,
+    fn: () => {
+      // The `jest.mock` transform will allow require, built-ins and globals.
+      const path = require('path');
+      const array = new Array(3);
+      array[0] = path.sep;
+      return jest.fn(() => array);
+    },
+  };
+});
 
 // These will not be hoisted
 jest.unmock('../__test_modules__/a').dontMock('../__test_modules__/b');
 jest.unmock('../__test_modules__/' + 'c');
 jest.dontMock('../__test_modules__/Mocked');
 
-
-describe('babel-plugin-jest-unmock', () => {
+describe('babel-plugin-jest-hoist', () => {
   it('hoists react unmock call before imports', () => {
     expect(typeof React).toEqual('object');
     expect(React.isValidElement.mock).toBe(undefined);
@@ -46,6 +63,27 @@ describe('babel-plugin-jest-unmock', () => {
 
     expect(d._isMockFunction).toBe(undefined);
     expect(d()).toEqual('unmocked');
+  });
+
+  it('hoists mock call with 2 arguments', () => {
+    const path = require('path');
+
+    expect(e._isMock).toBe(true);
+
+    const mockFn = e.fn();
+    expect(mockFn()).toEqual([path.sep, undefined, undefined]);
+  });
+
+  it('only executes the module factories once', () => {
+    global.CALLS = 0;
+
+    require('../__test_modules__/e');
+    expect(global.CALLS).toEqual(1);
+
+    require('../__test_modules__/e');
+    expect(global.CALLS).toEqual(1);
+
+    delete global.CALLS;
   });
 
   it('does not hoist dontMock calls before imports', () => {